+++ projects/sandbox/src/main.ts
@@ -1,7 +1,6 @@
 import {platformBrowser} from '@angular/platform-browser';
 import {enableProdMode, provideZoneChangeDetection} from '@angular/core';
 import {environment} from './environments/environment';
-// Application resources:
 import {AppModule} from './app/app.module';
 
 if (environment.production) {