+++ projects/sds/src/app/code-examples/code-examples/form/previews/input-prefixes-and-suffixes/form-example-input-prefixes-and-suffixes-preview.component.ts
@@ -9,10 +9,7 @@ import {ObErrorMessagesModule} from '@oblique/oblique';
 	selector: 'app-form-example-input-prefixes-and-suffixes-preview',
 	imports: [FormsModule, MatFormFieldModule, MatIconModule, MatInputModule, ObErrorMessagesModule],
 	templateUrl: './form-example-input-prefixes-and-suffixes-preview.component.html',
-	styleUrls: [
-		'./form-example-input-prefixes-and-suffixes-preview.component.scss',
-		'../../../../code-example-flex-layout.scss',
-	],
+	styleUrl: './form-example-input-prefixes-and-suffixes-preview.component.scss',
 })
 export class FormExampleInputPrefixesAndSuffixesPreviewComponent implements AfterViewInit {
 	/* The whole content of this class is simply a workaround for https://github.com/angular/components/issues/26428. This is an Angular bug