+++ projects/sds/src/app/code-examples/code-examples/spinner/spinner-code-examples.component.spec.ts
@@ -1,5 +1,6 @@
 import {type ComponentFixture, TestBed} from '@angular/core/testing';
 import {SpinnerCodeExamplesComponent} from './spinner-code-examples.component';
+import {provideObliqueTestingConfiguration} from '@oblique/oblique';
 
 describe('SpinnerCodeExamplesComponent', () => {
 	let component: SpinnerCodeExamplesComponent;
@@ -8,6 +9,7 @@ describe('SpinnerCodeExamplesComponent', () => {
 	beforeEach(async () => {
 		await TestBed.configureTestingModule({
 			imports: [SpinnerCodeExamplesComponent],
+			providers: [provideObliqueTestingConfiguration()],
 		}).compileComponents();
 
 		fixture = TestBed.createComponent(SpinnerCodeExamplesComponent);