+++ projects/sds/src/app/code-examples/code-examples/version/version-code-examples.component.ts
@@ -1,0 +1,27 @@
+import {CommonModule} from '@angular/common';
+import {ChangeDetectionStrategy, Component} from '@angular/core';
+import {IdPipe} from '../../../shared/id/id.pipe';
+import {CodeExampleComponent} from '../../code-example/code-example.component';
+import {type CodeExample, CodeExamples} from '../../code-examples.model';
+import {VersionExampleDefaultPreviewComponent} from './previews/default/version-example-default-preview.component';
+
+@Component({
+	selector: 'app-code-example-version',
+	imports: [CommonModule, IdPipe, CodeExampleComponent],
+	templateUrl: '../../code-examples.component.html',
+	changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class VersionCodeExamplesComponent extends CodeExamples {
+	readonly componentId = 'version-examples';
+	readonly previews: CodeExample[] = [
+		{
+			component: VersionExampleDefaultPreviewComponent,
+			idParts: ['default'],
+			title: 'Default',
+			snippets: [
+				this.getSnippet('version', 'default/version-example-default-preview.component.html', 'HTML'),
+				this.getSnippet('version', 'default/version-example-default-preview.component.ts', 'TS'),
+			],
+		},
+	];
+}